djangoapi/requirements.txt
Django==5.2
djangorestframework==3.16.0
drf-spectacular==0.28.0
gunicorn==23.0.0
python-dotenv==1.1.0
